feat 130: working on splitting the prefetcher
