--- src/Automatic/Contract/Exception/Exception.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Narrowspark\Automatic\Contract\Exception;
-
-interface Exception extends \Throwable
-{
-}
