+++ src/Common/Contract/Container.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Narrowspark\Automatic\Contract;
+namespace Narrowspark\Automatic\Common\Contract;
 
 interface Container
 {
@@ -11,7 +11,7 @@ interface Container
      *
      * @param string $id identifier of the entry to look for
      *
-     * @throws \Narrowspark\Automatic\Contract\Exception\InvalidArgumentException if no entry is found
+     * @throws \Narrowspark\Automatic\Common\Contract\Exception\InvalidArgumentException if no entry is found
      *
      * @return mixed
      */
@@ -26,6 +26,19 @@ interface Container
      * @return void
      */
     public function set(string $id, callable $callback): void;
+
+    /**
+     * Returns true if the container can return an entry for the given identifier.
+     * Returns false otherwise.
+     *
+     * `has($id)` returning true does not mean that `get($id)` will not throw an exception.
+     * It does however mean that `get($id)` will not throw a `NotFoundExceptionInterface`.
+     *
+     * @param string $id identifier of the entry to look for
+     *
+     * @return bool
+     */
+    public function has(string $id): bool;
 
     /**
      * Returns all container entries.
