+++ src/Prefetcher/.gitattributes
@@ -1,0 +1,4 @@
+* text=auto
+*.php text eol=lf
+
+README.md export-ignore
