+++ src/Prefetcher/Common/Contract/Exception/Exception.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Narrowspark\Automatic\Prefetcher\Common\Contract\Exception;
+
+/**
+ * This file is automatically generated, dont change this file, otherwise the changes are lost after the next mirror update.
+ *
+ * @codeCoverageIgnore
+ * @internal
+ */
+interface Exception extends \Throwable
+{
+}
