+++ src/Prefetcher/Common/Contract/Exception/InvalidArgumentException.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Narrowspark\Automatic\Contract\Exception;
+namespace Narrowspark\Automatic\Prefetcher\Common\Contract\Exception;
 
 use InvalidArgumentException as BaseInvalidArgumentException;
