+++ src/Prefetcher/Common/Contract/Exception/RuntimeException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Narrowspark\Automatic\Prefetcher\Common\Contract\Exception;
+
+use RuntimeException as BaseRuntimeException;
+
+final class RuntimeException extends BaseRuntimeException implements Exception
+{
+}
