+++ src/Prefetcher/CurlDownloader.php
@@ -80,7 +80,7 @@ final class CurlDownloader
     {
         $this->multiHandle = $mh = \curl_multi_init();
 
-        \curl_multi_setopt($mh, \CURLMOPT_PIPELINING, /*CURLPIPE_MULTIPLEX*/ 2);
+        \curl_multi_setopt($mh, \CURLMOPT_PIPELINING, /* CURLPIPE_MULTIPLEX */ 2);
 
         if (\defined('CURLMOPT_MAX_HOST_CONNECTIONS')) {
             \curl_multi_setopt($mh, \CURLMOPT_MAX_HOST_CONNECTIONS, 10);
