+++ src/Prefetcher/README.md
@@ -1,0 +1,49 @@
+<h1 align="center">Narrowspark Automatic</h1>
+<p align="center">
+    <a href="https://github.com/narrowspark/automatic/releases"><img src="https://img.shields.io/packagist/v/narrowspark/automatic.svg?style=flat-square"></a>
+    <a href="https://php.net/"><img src="https://img.shields.io/badge/php-%5E7.1.0-8892BF.svg?style=flat-square"></a>
+    <a href="https://codecov.io/gh/narrowspark/automatic"><img src="https://img.shields.io/codecov/c/github/narrowspark/automatic/master.svg?style=flat-square"></a>
+    <a href="#"><img src="https://img.shields.io/badge/style-level%207-brightgreen.svg?style=flat-square&label=phpstan"></a>
+    <a href="http://opensource.org/licenses/MIT"><img src="https://img.shields.io/badge/license-MIT-brightgreen.svg?style=flat-square"></a>
+</p>
+
+Branch Status
+------------
+[![Travis branch](https://img.shields.io/travis/narrowspark/automatic/master.svg?longCache=false&style=for-the-badge)](https://travis-ci.org/narrowspark/automatic)
+[![Appveyor branch](https://img.shields.io/appveyor/ci/narrowspark/automatic/master.svg?longCache=false&style=for-the-badge)](https://ci.appveyor.com/project/narrowspark/automatic/branch/master)
+
+> **Note:** This package is part of the [Narrowspark automatic](http://github.com/narrowspark/automatic). 
+
+Installation
+-------------
+
+Use [Composer](https://getcomposer.org/) to install this package:
+
+```sh
+composer require narrowspark/automatic-security-audit --dev
+```
+
+Usage
+-------------
+
+The checker will be executed when you launch `composer require` , `composer install` or `composer update`.
+If you have alerts in your composer.lock, `composer audit` will print them.
+
+Contributing
+------------
+
+Issues for this package shall be posted on [Narrowspark Automatic issues](http://github.com/narrowspark/automatic/issues). <br>
+Thank you for considering contributing to the Narrowspark automatic!
+
+> **Note:** Please note that this project is released with a Contributor Code of Conduct. By participating in this project you agree to abide by its terms.
+
+Credits
+-------------
+
+- [Daniel Bannert](https://github.com/prisis)
+- [All Contributors](../../contributors)
+
+License
+-------------
+
+The MIT License (MIT). Please see [License File](LICENSE) for more information.
