+++ src/Prefetcher/alias.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+\class_alias('Narrowspark\Automatic\Prefetcher\Common\Container', 'Narrowspark\Automatic\Common\Container');
+\class_alias('Narrowspark\Automatic\Prefetcher\Common\Contract\Container', 'Narrowspark\Automatic\Common\Contract\Container');
+\class_alias('Narrowspark\Automatic\Prefetcher\Common\Contract\Exception\Exception', 'Narrowspark\Automatic\Common\Contract\Exception\Exception');
+\class_alias('Narrowspark\Automatic\Prefetcher\Common\Contract\Exception\RuntimeException', 'Narrowspark\Automatic\Common\Contract\Exception\RuntimeException');
+\class_alias('Narrowspark\Automatic\Prefetcher\Common\Contract\Exception\InvalidArgumentException', 'Narrowspark\Automatic\Common\Contract\Exception\InvalidArgumentException');
+\class_alias('Narrowspark\Automatic\Prefetcher\Common\Contract\Resettable', 'Narrowspark\Automatic\Common\Contract\Resettable');
+\class_alias('Narrowspark\Automatic\Prefetcher\Common\Traits\GetGenericPropertyReaderTrait', 'Narrowspark\Automatic\Common\Traits\GetGenericPropertyReaderTrait');
+\class_alias('Narrowspark\Automatic\Prefetcher\Common\Traits\GetComposerVersionTrait', 'Narrowspark\Automatic\Common\Traits\GetComposerVersionTrait');
