+++ tests/Automatic/Fixture/testInitAutoScripts.json
@@ -1,0 +1,1 @@
+{"test":[]}
