+++ tests/Automatic/Fixture/testInitAutoScripts.lock
@@ -1,0 +1,1 @@
+{"packages":[]}
