+++ tests/Automatic/Fixture/testOnPostCreateProject.json
@@ -1,0 +1,1 @@
+{"name":"narrowspark\/narrowspark","type":"project","description":"A skeleton to start a new Narrowspark project.","license":"MIT"}
