+++ tests/Automatic/Fixture/testOnPostCreateProject.lock
@@ -1,0 +1,1 @@
+{"packages":[]}
