+++ tests/Automatic/composer_uninstall.lock
@@ -1,0 +1,1 @@
+[]
