+++ tests/Prefetcher/MirrorClassTest.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Narrowspark\Automatic\Prefetcher\Test;
+
+use PHPUnit\Framework\TestCase;
+
+/**
+ * @internal
+ *
+ * @small
+ */
+final class MirrorClassTest extends TestCase
+{
+}
